--- Public/test.html
@@ -1,73 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <meta http-equiv="X-UA-Compatible" content="IE=edge">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <link rel="stylesheet" href="https://cdn.jsdelivr.net/npm/bulma@0.9.3/css/bulma.min.css">
-    <title>Cookies for Coders</title>
-</head>
-<body>
-   <nav class="navbar has-shadow">
-       <div class="navbar-brand">
-           <a class="navbar-item">
-               <img src="Images/Cookies for coders.png" alt="logo for website" style="max-height: 65px" class="py-2 px-2">
-           </a>
-           <a class="navbar-burger" id="burger">
-               <span></span>
-               <span></span>
-               <span></span>
-           </a>
-       </div>
-       
-       <div class="navbar-menu" id="nav-links">
-            <div class="navbar-end">
-                <form method="get" action="insert_link_to_Account_page_here"><button type="submit"><a class="navbar-item">My Account</a></button></form>
-                <form method="get" action="insert_link_to_cart_here"><button type="submit"><a class="navbar-item">Cart (0)</a></button></form>
-            </div>
-        </div>
-
-
-    </nav>
-
-<!--breadcrumb section-->
-    <div class="section pt-4 pb-0">
-        <nav class="breadcrumb">
-            <ul class="container">
-                <form method="get" action="insert_link_to_homepage_here"><button type="submit"><a class="has-text-black">Cookies for Coders</a></button></form> <hr>
-                <form method="get" action="insert_link_to_menu_here"><button type="submit"><a class="has-text-black">Shop</a></button></form> <hr>
-                <form method="get" action="insert_link_to_???_here"><button type="submit"><a>Cookies</a></button></form>
-            </ul>
-        </nav>
-    </div>
-
-
-
-    <!--Info for products-->
-    <section class="section">
-        <div class="container">
-            <div class="columns is-vcentered">
-                <div class="column is-3">
-                    <h1 class="is-size-4 title has-text-centered">Cookies for Coders</h1>
-                    <h2 class="is-size-6 subtitle has-text-centered">Cookies</h2>
-                    <p class="has-text-centered">Cookies for those ALL NIGHTERS!!!</p>
-                </div>
-
-                <div class="column is-6 has-text-centered">
-                    <img src="Images/chocolate-chip-cookie-16.png" alt="pic of cookie" style="max-height: 500px;" class="px-6">
-                </div>
-
-
-                <div class="column is-3 has-text-centered">
-                    <div class="is-size-3 mb-4">$2.99</div>
-                    <p>The best three dollars you've ever spent! Trust us!</p>
-                </div>
-                
-            </div>
-        </div>
-    </section>
-   
-   
-    <script src="assets/js/index.js"></script>
-</body>
-</html>
