--- Wireframes
@@ -1,1 +1,0 @@
-Hello everyone. Excited to be working on this project with you all!
