--- controllers/index.js
@@ -1,7 +1,0 @@
-const router = require('express').Router();
-
-const primaryRoutes = require('./primaryRoutes');
-
-router.use('/', primaryRoutes);
-
-module.exports = router;
